Allmost finalise transition to new coloring
Colors from field.lua moved into palette.lua. But.
There is some colors that I called Mystic_*. It is like that,
because as a contributor I wasn't been able to figure out what are
this colors actually for.
I'll ask a head developer of this project about these colors, so
I will be able to rename them properly